--- README.md
@@ -1,56 +1,0 @@
-# Friend_Builder
-
-
-# Project Description:
-Android app to help people who have similar interests connect and chat: Use GPS to locate and match similar interests (alternatively, have the user set up the info themselves and allow them to read the descriptions of other users in the vicinity)
-
-
-## MVP (Minimal Feature Set):
-
-
-User account
-
-User profile
-
-Messaging between users
-
-GPS based distance between users visible
-
-Users create events for others to see
-
-Users can hide from/ignore other users they don't like
-
-
-## Add-on Features (if we get around to them):
-
-
-App gives recommendations to users based on user supplied interests
-
-Users can send pictures via messaging
-
-Users interactions with each other generate "experience points" with each other, which unlocks features
-
-Spot a stranger feature
- 
-
-## User Stories
-
-I am a student new to a large university. This school is much larger than my high school and the community college I had attended previously and I'm finding the giant class sizes a little bit overwhelming. It was easy to make friends when there were only a few of us in the classroom, but it's much more difficult to make friends here since everyone seems to be so different. Fortunately for me, I had heard about this app called Friend_Builder that was designed specifically for people like me who might be a little bit shy at first. I simply download the app to my android phone, set up my profile by using my e-mail address and a login password. After I login I am prompted to answer a few questions about my interests. I am then connected to a bunch of people nearby with similar interests to mine. I can send messages to people to see if they'd be interested in meeting up for coffee or lunch. I can also rate people based on whether or not we hit it off or not. Overall I'm hoping that this app will allow me to connect with people who share the same academic interests and hobbies with me.
-
-
-“I want to meet people near me who share my interests.”
-
-“I want to find events to attend with people who share that interest.”
-
-  -Example: “Sometimes I need a gym buddy but feel uncomfortable walking up to random people.”
-  
-  -Example: “Sometimes I want to play basketball but I don’t know if anyone is willing to play.”
-  
-“I want a way to keep in touch with those people, if I like them, and a way to cut contact with them, if I don’t like them.”
-
-
-## Gitter
-
-Please feel free to contact members of the team on Gitter by following the badge below:
-
-[![Gitter chat](https://badges.gitter.im/gitterHQ/gitter.png)](https://gitter.im/Friend_Builder/Lobby?utm_source=ios&utm_medium=link&utm_campaign=ios-share-link)
