--- app/src/main/java/edu/buffalo/friendbuilder/Database.java
@@ -1,15 +1,0 @@
-package edu.buffalo.friendbuilder;
-
-import java.util.HashMap;
-
-/**
- * Created by PeterYang on 2/16/2018.
- */
-
-public class Database{
-    private HashMap<String, String> data = new HashMap<>();
-
-    public java.util.HashMap<String, String> getMap(){
-        return data;
-    }
-}
